Refactor code in scrollable container component
In this version, I've:

Added default values for the projects, blogs, and workHistory props so that the component doesn't break if they're not passed in.
Used the ternary operator to conditionally render the renderProjects, renderBlogs, and renderWorkHistory variables.
Removed the curly braces around the JSX inside the return statement since they're not necessary when rendering a single element.
Added self-closing tags to the Subtitle, ProjectItem, BlogItem, Skill, and NavigationButton components. This is optional but considered good practice.